+++ src/fintrist_ds/engine.py
@@ -46,7 +46,7 @@ def store_result(name, process, parents=None, params=None, **kwargs):
     """Use a local or library function to create and run a new Study."""
     if isinstance(process, str):
         process = get_function(process)
-    newstudy = create_study(name, process, parents, params, **kwargs)
+    newstudy = create_study(name, process, parents, params=params, **kwargs)
     newstudy.run(function=process)
     newstudy.save()
     return newstudy